example: move handler to child dir to test file watching
