+++ examples/okpc/handlers/index.ts
@@ -1,7 +1,9 @@
-import { CanonicalTransactionChain } from "./CanonicalTransactionChain";
+import { CanonicalTransactionChain } from "./dir/CanonicalTransactionChain";
 import { OKPC } from "./OKPC";
 
 export default {
   OKPC,
   CanonicalTransactionChain,
 };
+
+// haha
